Removed all "static" variables from IncomingHandlerInterface
As you can see, the static variables weren't required as such, except the BattleShipGUI
	You can rewrite the code such that the BattleShipGUI doesn't have to be static, as well, but I don't consider this to be a "bad" use of a static - in this case, you're never instantiating the IncomingHandlerInterface, so using that as a static variable could be a valid use
	Note that one thing that could be done here to help enforce that rule is to create that default constructor, but make it private (as I discussed in the previous commit) - "private IncomingHandlerInterface()"

Understand that the reason we want as few static variables as possible is because it creates "global" state, which is inherently unsafe. If any of the code in your application, it becomes even more unsafe. "static" code should be used sparingly. If you have questions about statics in particular, let me know.