Continue refactoring instance variables
Removed many of the class level variables that weren't required - All of the JLabels were only being referenced in one place (the initialization method), so we don't have to have these at class scope. This means that the Graphics framework can maintain them and their references, and that they don't clutter up the class interface when they're not required. Clean code AND possible better performance - win/win!

Changed instructions to a static final INSTRUCTIONS variable - this approach, or putting it directly into the code inline (eg "instructionTextArea.append(<text here>...)" is preferable to having it in just a "private String instructions", because it is a constant and should be treated as such. However, it would be better still to put this into some kind of properties, configuration or internationalization text file so that you can just reference that file from the code. As a text file, you could then replace "Instructions\nEnter your username...." with "Anweisungen\nTrage deinen Namen in den Box ein..." (which is German) or whatever other language you wanted to. This is a far more widely accepted practice. If you'd like an example, we can walk through one, but it's more in the scope of 4230.